Experiment with providing thread builder into the workers
Problem currently is that thread::spawn needs to take ownership of the builder. Since you can't take the value out of Mutex easily, we need a better solution. Even if we could, then the problem is after the first thread takes the builder, other threads can never.

Possible solution is to implement a copyable thread_builder