Restructure the project into a single crate
There was no need to make a new crate for every module. My initial reasoning was to make the project as modular as possible. I think most of the functionality is small enough that a single crate would be fine for now.

It is definitely possible threadpool may become a seperate crate in the future. But it is fine as it is.